Systemd maldet.service should source a value for default_monitor_mode.
The current maldet.service script always starts maldet in monitor mode using
the monitor_paths file. Let's set the monitor mode based on the value for
$default_monitor_mode sourced from conf.maldet.

I don't know if there's a way to do the same IF structure found in maldet.sh in
a systemd unit. We're stuck with relying solely on conf.maldet, since
/etc/sysconfig/maldet is setting a value for a different variable,
$MONITOR_MODE.

Resource:
https://coreos.com/os/docs/latest/using-environment-variables-in-systemd-units.html